+++ src/main/groovy/rocks/metaldetector/butler/web/api/ReleasesResponse.groovy
@@ -1,6 +1,7 @@
-package rocks.metaldetector.butler.web.dto
+package rocks.metaldetector.butler.web.api
 
 import groovy.transform.Canonical
+import rocks.metaldetector.butler.web.dto.ReleaseDto
 
 @Canonical
 class ReleasesResponse {
