+++ src/main/groovy/rocks/metaldetector/butler/web/rest/ImportJobRestController.groovy
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RestController
 import rocks.metaldetector.butler.service.importjob.ImportJobService
+import rocks.metaldetector.butler.web.api.ImportJobResponse
 import rocks.metaldetector.butler.web.dto.ImportJobDto
-import rocks.metaldetector.butler.web.dto.ImportJobResponse
 
 import static rocks.metaldetector.butler.config.constants.Endpoints.COVER_JOB
 import static rocks.metaldetector.butler.config.constants.Endpoints.IMPORT_JOB
